Refactor of mapViewer.
