+++ examples/mapViewer/source/asset/map.c
