+++ examples/mapViewer/source/asset/map.h
