+++ examples/mapViewer/source/mapViewer/main.c
@@ -1,5 +1,6 @@
 #include "../gbaLib/types.h"
 #include "mapViewer.h"
+#include "map/map.h"
 #include "../gbaLib/input/input.h"
 
 int main()
@@ -17,7 +18,6 @@ int main()
         shift.y += getYAxis(inputState) * SHIFT_SPEED;
         shift.x += getXAxis(inputState) * SHIFT_SPEED;
 
-        for (u16 layer = 0; layer < map.numLayers; layer++)
-            shiftMap(layer, shift);
+        shiftMap(map, shift);
     }
 }
